requirements.txt
fastapi 
uvicorn 
python-multipart 
pydantic-setting
yt-dlp
openai-whisper
